Begin the PlayGame function
I started working on the playgame function here. I left a TODO @ line 44, but I think it will be better (and possibly even necessary) to make the code a bit more functionally focuses. I am going to work on that later, but if anyone else wants to take a crack at it, feel free!